requirements.txt
sanic==22.6.2
# transformers
accelerate
onnx
onnxruntime
numpy
Pillow
requests